refactor: more utils
