+++ modules/resources/components/submit-resource-welcome-card.tsx
@@ -1,7 +1,7 @@
 import { currentUser } from '@clerk/nextjs';
 
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/utils/cn';
 
 type SubmitResourceWelcomeCardProps = {
   className?: string;
